feat: Integrate multi component widget UI

+++ features/merchant/seller_home_common/src/main/java/com/tokopedia/sellerhomecommon/presentation/view/viewholder/BarChartViewHolder.kt
@@ -253,7 +253,7 @@ class BarChartViewHolder(
                 barHexColor = it.barHexColor,
                 values = it.value.mapIndexed { i, item ->
                     val xLabel = getXLabel(xLabels, i)
-                    return@mapIndexed BarChartMetricValue(item.value, item.valueFmt, xLabel)
+                    return@mapIndexed BarChartMetricValue(item.value.toFloat(), item.valueFmt, xLabel)
                 }
             )
         }.orEmpty()
